% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loo_cv.R
\name{get_k_step_ahead_bw}
\alias{get_k_step_ahead_bw}
\title{get_k_step_ahead_bw: The my happy function}
\usage{
get_k_step_ahead_bw(data, nneighbor, week, h, num_lags)
}
\arguments{
\item{x}{numeric number}

\item{...}{other arguments}
}
\description{
get_k_step_ahead_bw: The my happy function
}
