% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{periodic}
\alias{periodic}
\title{Periodic Kernel 
Inputs:
x = variable 1
y = variable 2
h = inverse of kernel bandwith
rho = periodicity of kernel
eta = smoothing parameter
Outputs:
similarity between x and y}
\usage{
periodic(x, y, h = 1, rho = pi/52, eta = 1e+10)
}
\description{
Periodic Kernel 
Inputs:
x = variable 1
y = variable 2
h = inverse of kernel bandwith
rho = periodicity of kernel
eta = smoothing parameter
Outputs:
similarity between x and y
}
