% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loo_cv.R
\docType{data}
\name{possible_sigmas}
\alias{possible_sigmas}
\title{Leave one season out cross validation to estimate kernel parameters
Inputs:
ts_obj = time series object containing a time series under 
             ts_obj$val
         and a season indicator under
              ts_obj$season
         and a calendar date under
              ts_obj$date
k = number of neighbors to use
h = horizon 
num_lags = 4
sigma = bandwith
Outputs:}
\format{An object of class \code{numeric} of length 11.}
\usage{
possible_sigmas
}
\description{
Leave one season out cross validation to estimate kernel parameters
Inputs:
ts_obj = time series object containing a time series under 
             ts_obj$val
         and a season indicator under
              ts_obj$season
         and a calendar date under
              ts_obj$date
k = number of neighbors to use
h = horizon 
num_lags = 4
sigma = bandwith
Outputs:
}
\keyword{datasets}
