% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{rbf}
\alias{rbf}
\title{Set of Kernel Functions
Currently implemented 
1) Gaussian
2) periodic}
\usage{
rbf(x, y, h = 1)
}
\description{
Gussian Kernel 
Inputs:
x = variable 1
y = variable 2
h = inverse of kernel bandwith
Outputs:
similarity between x and y
}
